Add receive command
The receive command allows to receive snapshot data from stdin and store
it in a zfs file system. Care must be taken that the target file system
is excluded when creating snapshots.

Additionally this commit restructures the tests for the zfs wrapper. The
new structure is more generic and makes it easier to write additional
tests.